Initial commit
The simple GPS app tries to show local navigation in a circled manner,
where the range of the circle shows to approximation of the points to
user current location, and the colour shows the distance between. The
nearer of the target, the colour darkens, and the further the distance,
the colour lightens.